backend/pyproject.toml
[project]
name = "pawcord-backend"
version = "0.1.0"
description = "Pawcord FastAPI backend"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.30",
  "alembic>=1.14",
  "pydantic-settings>=2.0",
  "pydantic[email]>=2.0",
  "python-jose[cryptography]>=3.3",
  "argon2-cffi>=23.1",
  "pynacl>=1.5",
  "redis[hiredis]>=5.2",
  "arq>=0.25",
  "slowapi>=0.1.9",
  "python-multipart>=0.0.20",
  "minio>=7.2",
  "python-magic>=0.4",
  "pillow>=10.4",
  "aiofiles>=24.1",
  "pytest-asyncio>=0.24",
  "httpx>=0.28",
  "factory-boy>=3.3",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]
